requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.9

django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-microsoft-auth==3.0.1
django-microsoft-authentication==0.1.6
idna==3.6
msal==1.26.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-oauth2==1.1.1
python-social-auth==0.3.6
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
